コントローラリファクタリングに伴うテストの修正
